\name{ccme_wqi-package}
\alias{ccme_wqi-package}
\alias{ccme_wqi}
\docType{package}
\title{
\packageTitle{ccme_wqi}
}
\description{
\packageDescription{ccme_wqi}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ccme_wqi}
\packageIndices{ccme_wqi}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ccme_wqi}

Maintainer: \packageMaintainer{ccme_wqi}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
