\name{ccmewqi-package}
\alias{ccmewqi-package}
\alias{ccmewqi}
\docType{package}
\title{
\packageTitle{ccmewqi}
}
\description{
\packageDescription{ccmewqi}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ccmewqi}
\packageIndices{ccmewqi}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ccmewqi}

Maintainer: \packageMaintainer{ccmewqi}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
