% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/excursion.R
\name{excursion}
\alias{excursion}
\title{excursion function}
\usage{
excursion(exc, test_value, objective_one, objective_two, limit = c("<", ">",
  "<>"))
}
\arguments{
\item{exc}{}

\item{test_value}{}

\item{objective_one}{}

\item{objective_two}{}

\item{limit}{}
}
\description{
Calculate the magnitude of objective failures.
}
\details{
The number of times by which an individual concentration is greater than
(or less than, when the objective is a minimum) the objective is termed
an "excursion"
}
\seealso{
\url{http://www.ccme.ca/en/resources/canadian_environmental_quality_guidelines/calculators.html}
}

