% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1.R
\name{F1}
\alias{F1}
\title{F1 (Scope) function.}
\usage{
F1(a, b)
}
\arguments{
\item{a}{Failed variables}

\item{b}{Total variables}
}
\description{
Calculates the percentage of variables that do not meet their
objectives at least once during the time period under consideration
("failed variables"), relative to the total number of variables
measured:
}
\details{
Inputs must be numeric.
\deqn{F_{1} = \frac{Number of failed variables}{Total number of variables} * 100}
}
\seealso{
\url{http://www.ccme.ca/en/resources/canadian_environmental_quality_guidelines/calculators.html}
}

