% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/F2.R
\name{F2}
\alias{F2}
\title{F2 (Frequency) function.}
\usage{
F2(a, b)
}
\arguments{
\item{a}{Failed tests}

\item{b}{Total tests}
}
\description{
Represents the percentage of individual tests that do not meet meet objectives ("failed tests"):
}
\details{
Inputs must be numeric.
\deqn{F_{2} = \frac{Number of failed tests}{Total number of tests} * 100}
}
\seealso{
\url{http://www.ccme.ca/en/resources/canadian_environmental_quality_guidelines/calculators.html}
}

