% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F3.R
\name{F3}
\alias{F3}
\title{F3 (Amplitude) function.}
\usage{
F3(nse)
}
\arguments{
\item{nse}{normalized sum of excursions}
}
\description{
Represents the amount by which failed test values do not meet their objectives
}
\details{
Inputs must be numeric.
\deqn{F_{3} = \frac{Normalized sum of excursions}{0.01 * normalized sum of excursions + 0.01}}
}
\seealso{
\url{http://www.ccme.ca/en/resources/canadian_environmental_quality_guidelines/calculators.html}
}
