% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_exceedance.R
\name{is_exceedance}
\alias{is_exceedance}
\title{is_exceedance function.}
\usage{
is_exceedance(test_value, lower_limit = NA_real_, upper_limit = NA_real_,
  limit = c("<", ">", "<>"))
}
\arguments{
\item{limit}{}
}
\value{
boolean True if test value exceeded objective
}
\description{
Check if test value fall outside of acceptable objective limit.
}
\details{
Inputs must be numeric.
}
\seealso{
\url{http://www.ccme.ca/en/resources/canadian_environmental_quality_guidelines/calculators.html}
}
