% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nse.R
\name{nse}
\alias{nse}
\title{nse function.}
\usage{
nse(a, b)
}
\arguments{
\item{a}{Sum of excursions}

\item{b}{Total number of tests}
}
\description{
Normalized sum of excursions from objectives
}
\details{
Inputs must be numeric.
}
\seealso{
\url{http://www.ccme.ca/en/resources/canadian_environmental_quality_guidelines/calculators.html}
}

