% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion.R
\name{excursion}
\alias{excursion}
\title{excursion function}
\usage{
excursion(test_value, lower_limit = NA_real_, upper_limit = NA_real_)
}
\arguments{
\item{test_value}{}
}
\description{
Calculate the magnitude of objective failures.
}
\details{
The number of times by which an individual concentration is greater than
(or less than, when the objective is a minimum) the objective is termed
an "excursion"
}
\seealso{
\url{http://www.ccme.ca/en/resources/canadian_environmental_quality_guidelines/calculators.html}
}
