% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is_exceedance.R
\name{is_exceedance}
\alias{is_exceedance}
\title{is_exceedance function.}
\usage{
is_exceedance(test_value, objective_one, objective_two, limit = c("<", ">",
  "<>"))
}
\arguments{
\item{test_value}{}

\item{objective_one}{}

\item{objective_two}{}

\item{limit}{}
}
\value{
boolean True if test value exceeded objective
}
\description{
Check if test value fall outside of acceptable objective limit.
}
\details{
Inputs must be numeric.
}
\seealso{
\url{http://www.ccme.ca/en/resources/canadian_environmental_quality_guidelines/calculators.html}
}

