% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqi.R
\name{wqi}
\alias{wqi}
\title{wqi function.}
\usage{
wqi(F1, F2, F3, method = NULL)
}
\arguments{
\item{method}{optional CCAMP method.}
}
\description{
Final CCME water quality index score
}
\details{
Inputs must be numeric.
}
\seealso{
\url{http://www.ccme.ca/en/resources/canadian_environmental_quality_guidelines/calculators.html}
}
